schemes-chatbot/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
langchain==0.3.27
langchain-groq==0.3.6
langchain-core==0.3.72
pydantic==2.11.7
python-multipart==0.0.12
asyncpg==0.30.0
aiohttp==3.10.10
beautifulsoup4==4.12.3
requests==2.32.3
pdfminer.six==20221105
